requirements.txt
pandas==1.4.1
psycopg2==2.9.3
pysftp==0.2.9
python-dotenv==0.19.2
pandas-schema==0.3.5